remove definitions.ts
