+++ src/commands.ts
@@ -1,10 +1,8 @@
 import {AtCoder} from "./atcoder";
 import {OnlineJudge} from "./facade/oj";
 import {Cookie} from "./cookie";
-import * as project from "./project";
-import {Contest, Task} from "./definitions";
+import {Contest, Task, detectTaskByPath, findProjectJSON, formatTaskDirname, saveProjectJSON, init, installTask} from "./project";
 import getConfig, {defaults, getConfigDirectory} from "./config";
-import {formatTaskDirname, saveProjectJSON} from "./project";
 import {getTemplate, getTemplates} from "./template";
 
 export async function login() {
@@ -30,7 +28,7 @@ export async function contest(contest_id: string | undefined, options: { id?: bo
 	if (contest_id === undefined) {
 		// idが与えられていない場合、プロジェクトファイルを探してコンテスト情報を表示
 		try {
-			const {data: {contest}} = await project.findProjectJSON();
+			const {data: {contest}} = await findProjectJSON();
 			console.log(format(contest));
 		} catch (e) {
 			console.error(e.message);
@@ -54,7 +52,7 @@ export async function task(contest_id: string | undefined, task_id: string | und
 	if (contest_id === undefined && task_id === undefined) {
 		// idが与えられていない場合、プロジェクトファイルを探す
 		try {
-			const {task} = await project.detectTaskByPath();
+			const {task} = await detectTaskByPath();
 			if (task === null) {
 				console.error("failed to find the task.");
 				return;
@@ -85,7 +83,7 @@ export async function tasks(contest_id: string | undefined, options: { id?: bool
 	if (contest_id === undefined) {
 		// idが与えられていない場合、プロジェクトファイルを探す
 		try {
-			const {data: {tasks}} = await project.findProjectJSON();
+			const {data: {tasks}} = await findProjectJSON();
 			console.log(format(tasks));
 		} catch (e) {
 			console.error(e.message);
@@ -154,7 +152,7 @@ export async function submit(filename: string, options: { task?: string, contest
 	let task_id = options.task;
 	if (contest_id === undefined || task_id === undefined) {
 		// コンテストまたはタスクが未指定の場合、カレントディレクトリのパスから提出先を調べる
-		const {contest, task} = await project.detectTaskByPath();
+		const {contest, task} = await detectTaskByPath();
 		if (contest_id === undefined && contest !== null) contest_id = contest.id;
 		if (task_id === undefined && task !== null) task_id = task.id;
 
@@ -259,7 +257,7 @@ function undef2empty(strings: TemplateStringsArray, ...values: Array<any>): stri
 
 export async function setup(contest_id: string, options: { choice: "inquire" | "all" | "none" | "rest" | "next", force?: boolean, contestDirnameFormat?: string, taskDirnameFormat?: string, template?: string }) {
 	try {
-		const {contest} = await project.init(contest_id, options);
+		const {contest} = await init(contest_id, options);
 		console.log(`create project of ${contest.title}`);
 		await add(options);
 	} catch (e) {
@@ -269,7 +267,7 @@ export async function setup(contest_id: string, options: { choice: "inquire" | "
 
 export async function add(options: { choice: "inquire" | "all" | "none" | "rest" | "next", force?: boolean, taskDirnameFormat?: string, template?: string }) {
 	try {
-		const {path, data} = await project.findProjectJSON();
+		const {path, data} = await findProjectJSON();
 		const {contest, tasks} = data;
 		const choices = await selectTasks(tasks, options.choice, options.force);
 		const template = options.template !== undefined ? (await getTemplate(options.template).catch((e) => {
@@ -281,7 +279,7 @@ export async function add(options: { choice: "inquire" | "all" | "none" | "rest"
 			const format = options.taskDirnameFormat !== undefined ? options.taskDirnameFormat : (await getConfig()).get("default-task-dirname-format");
 			const dirname = formatTaskDirname(format, task, index, contest);
 			// 新しいTaskが返ってくるので、もともとの配列の要素を更新する
-			tasks[index] = await project.installTask(task, index, contest, dirname, path, template);
+			tasks[index] = await installTask(task, index, contest, dirname, path, template);
 		}
 		// 更新されたContestProjectをファイルに書き出し
 		await saveProjectJSON(Object.assign(data, {tasks}));
